requirements2.txt
asgiref==3.6.0
beautifulsoup4==4.12.2
contourpy==1.0.7
crispy-bootstrap4==2022.1
cycler==0.11.0
Django==4.2.1
django-bootstrap4==23.1
django-crispy-forms==2.0
fonttools==4.39.4
kiwisolver==1.4.4
matplotlib==3.7.1
numpy==1.24.3
packaging==23.1
pandas==2.0.2
patsy==0.5.3
Pillow==9.5.0
plotly==5.14.1
plotly-express==0.4.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
scipy==1.10.1
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
statsmodels==0.14.0
tenacity==8.2.2
tzdata==2023.3
